refactor: applying name convention to folder name
