+++ packages/sdk/src/search/globalState/useSearchState.ts
